Update x-approver bug on changing approver and has the same approver id

+++ lvis-api/apps/warehouse/src/gas-slip-approver/gas-slip-approver.service.ts
@@ -29,6 +29,11 @@ export class GasSlipApproverService {
                     gas_slip: {
                         select: {
                             gas_slip_number: true,
+                            gas_slip_approvers: {
+                                orderBy: {
+                                    order: 'asc'
+                                }
+                            }
                         },
                     },
                 },
@@ -42,13 +47,6 @@ export class GasSlipApproverService {
                 throw new BadRequestException('Can only change approver if status is pending')
             }
 
-            const updateApprover = await prisma.gasSlipApprover.update({
-                where: { id },
-                data: {
-                    approver_id: input.new_approver_id,
-                },
-            });
-
             const pending = await prisma.pending.findUnique({
                 where: {
                     approver_id_reference_number_reference_table: {
@@ -61,21 +59,57 @@ export class GasSlipApproverService {
 
             if (pending) {
 
-                // delete previous approver's pending
-                await prisma.pending.delete({
-                    where: { id: pending.id },
-                });
+                const approvers = item.gas_slip.gas_slip_approvers.filter(i => i.approver_id === item.approver_id)
 
-                // add pending for new approver
-                await prisma.pending.create({
-                    data: {
-                        approver_id: input.new_approver_id,
-                        reference_number: pending.reference_number,
-                        reference_table: pending.reference_table,
-                        description: pending.description,
-                    },
-                });
-            } 
+                // check if approver to update has duplicates
+                if(approvers.length > 1) {
+
+                    const leastOrder = approvers.reduce((min, obj) => {
+                        return obj.order < min.order ? obj : min
+                    }, approvers[0])
+
+                    // can update since approver to update has the least order. Meaning first on the queue
+                    if(item.id === leastOrder.id) {
+
+                        // delete previous approver's pending
+                        await prisma.pending.delete({
+                            where: { id: pending.id },
+                        });
+        
+                        // add pending for new approver
+                        await prisma.pending.create({
+                            data: {
+                                approver_id: input.new_approver_id,
+                                reference_number: pending.reference_number,
+                                reference_table: pending.reference_table,
+                                description: pending.description,
+                            },
+                        });
+                    }
+                } else {
+                    // delete previous approver's pending
+                    await prisma.pending.delete({
+                        where: { id: pending.id },
+                    });
+    
+                    // add pending for new approver
+                    await prisma.pending.create({
+                        data: {
+                            approver_id: input.new_approver_id,
+                            reference_number: pending.reference_number,
+                            reference_table: pending.reference_table,
+                            description: pending.description,
+                        },
+                    });
+                }
+            }
+
+            const updateApprover = await prisma.gasSlipApprover.update({
+                where: { id },
+                data: {
+                    approver_id: input.new_approver_id,
+                },
+            });
 
             return updateApprover;
         });
